Portail_CDR_Backend/requirements.txt
aiodns==3.0.0
aiosmtpd==1.4.2
asgiref==3.4.1
atpublic==2.3
attrs==22.2.0
cffi==1.15.1
Django==3.2.21
django-auth-ldap==4.0.0
django-cors-headers==3.10.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.1.0
openpyxl==3.1.2
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycares==4.3.0
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-ldap==3.4.3
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.1.1
tzdata==2023.3
verify-email==2.4.3
xlwt==1.3.0
